+++ src/core/kmerhash.cpp
@@ -103,7 +103,7 @@ inline auto kh_insert_kmer(struct kh_handle_s & kmer_hash,
 
 auto kh_insert_kmers(struct kh_handle_s & kmer_hash, int const k_offset, char const * seq, int const len) -> void
 {
-  int const kmers = 1U << (2U * static_cast<unsigned int>(k_offset));
+  int const kmers = static_cast<int>(1U << (2U * static_cast<unsigned int>(k_offset)));
   auto const kmer_mask = static_cast<unsigned int>(kmers - 1);
 
   reset_buckets(kmer_hash.hash);
@@ -165,7 +165,7 @@ auto kh_find_diagonals(struct kh_handle_s const & kmer_hash,
                        std::vector<int> & diags) -> void
 {
 
-  int const kmers = 1U << (2U * static_cast<unsigned int>(k_offset));
+  int const kmers = static_cast<int>(1U << (2U * static_cast<unsigned int>(k_offset)));
   auto const kmer_mask = static_cast<unsigned int>(kmers - 1);
 
   unsigned int bad = kmer_mask;
